refactor(mediastream): Moves the isKeyFrame method to VideoMediaStream.
