+++ src/org/jitsi/impl/neomedia/rtp/translator/RTCPFeedbackMessageSender.java
@@ -20,6 +20,7 @@ import java.util.concurrent.atomic.*;
 
 import org.jitsi.impl.neomedia.*;
 import org.jitsi.impl.neomedia.rtp.*;
+import org.jitsi.service.neomedia.*;
 import org.jitsi.service.neomedia.event.*;
 import org.jitsi.util.*;
 import org.jitsi.util.concurrent.*;
@@ -268,8 +269,8 @@ public class RTCPFeedbackMessageSender
                 return;
             }
 
-            if(!streamRTPManager
-                .streamRTPManager.getMediaStream().isKeyFrame(buf, off, len))
+            if(!((VideoMediaStream)streamRTPManager
+                .streamRTPManager.getMediaStream()).isKeyFrame(buf, off, len))
             {
                 return;
             }
