+++ src/org/jitsi/service/neomedia/VideoMediaStream.java
@@ -156,4 +156,14 @@ public interface VideoMediaStream
      * {@link MediaStream} if one doesn't already exist. Returns the instance.
      */
     public BandwidthEstimator getOrCreateBandwidthEstimator();
+
+    /**
+     * Utility method that determines whether or not a packet is a key frame.
+     *
+     * @param buf the buffer that holds the RTP payload.
+     * @param off the offset in the buff where the RTP payload is found.
+     * @param len then length of the RTP payload in the buffer.
+     * @return true if the packet is a key frame, false otherwise.
+     */
+    public boolean isKeyFrame(byte[] buf, int off, int len);
 }
